+++ resources/js/builder/inventory/InventoryBuilder.tsx
@@ -6,10 +6,17 @@ import api from "../services/api"
 import ButtonConfiguration from "./components/configuration/button/ButtonConfiguration";
 import ItemStackConfiguration from "./components/configuration/itemstack/ItemStackConfiguration";
 
+const ITEMS_PER_PAGE = 54;
+const MAX_PAGE = 10;
+// The page management system is temporary, I have not yet found a better solution than to generate the whole table in advance.
+// We have to think about a better solution because the multi page and else button system will not be compatible in the state
+
 const InventoryBuilder = () => {
 
     const findButton = (index) => {
-        return inventory.buttons.find(button => button.slot == index) || null;
+        let page = Math.ceil((index + 1) / ITEMS_PER_PAGE)
+        let currentSlot = index - (ITEMS_PER_PAGE * (page - 1))
+        return inventory.buttons.find(button => button.slot == currentSlot && button.page == page) || null;
     }
 
     // @ts-ignore
@@ -20,15 +27,20 @@ const InventoryBuilder = () => {
     const [isShiftClick, setIsShiftClick] = useState(false);
     const [needToUpdate, setNeedToUpdate] = useState(false);
     const [slots, setSlots] = useState([]);
+    const [page, setPage] = useState(1);
     const [inventoryContent, setInventoryContent] = useState({
         currentSlot: 0,
         // @ts-ignore
-        slots: Array.from({length: 54}, (_, index) => {
+        slots: Array.from({length: ITEMS_PER_PAGE * MAX_PAGE}, (_, index) => {
             let button = findButton(index)
+            let page = Math.ceil((index + 1) / ITEMS_PER_PAGE)
+            let currentSlot = index - (ITEMS_PER_PAGE * (page - 1))
             return ({
                 id: index,
                 content: button?.item ?? null,
                 button: button ?? {
+                    slot: currentSlot,
+                    page: page,
                     model_id: 0,
                     amount: 0,
                     display_name: null,
@@ -44,7 +56,9 @@ const InventoryBuilder = () => {
         })
     });
 
+
     useEffect(() => {
+        console.log(inventoryContent)
         const intervalId = setInterval(() => {
             saveData();
         }, 1000 * 30); // 30 seconds
@@ -274,6 +288,7 @@ const InventoryBuilder = () => {
     }, [currentItem])
 
     const updateSlotContent = (slotIndex, newContent, newAmount = 1) => {
+        slotIndex = parseInt(slotIndex) + ((page - 1) * ITEMS_PER_PAGE)
         setNeedToUpdate(true)
         setInventoryContent(prevInventoryContent => {
             const newSlots = [...prevInventoryContent.slots];
@@ -429,7 +444,8 @@ const InventoryBuilder = () => {
             if (slot.content != null) {
                 formData.append(`slot[${index}]item_id`, slot.content.id);
                 formData.append(`slot[${index}]amount`, slot.button.amount);
-                formData.append(`slot[${index}]slot`, slot.id);
+                formData.append(`slot[${index}]slot`, slot.button.slot);
+                formData.append(`slot[${index}]page`, slot.button.page);
                 formData.append(`slot[${index}]name`, slot.button.name);
                 formData.append(`slot[${index}]is_permanent`, slot.button.is_permanent);
                 formData.append(`slot[${index}]close_inventory`, slot.button.close_inventory);
@@ -487,7 +503,8 @@ const InventoryBuilder = () => {
                     <Inventory inventory={inventory} updateInventory={updateInventory}
                                inventoryContent={inventoryContent}
                                needToUpdate={needToUpdate} saveData={saveData} selectSlots={slots}
-                               handleSlotClick={handleSlotClick} handleSlotDoubleClick={handleSlotDoubleClick}/>
+                               handleSlotClick={handleSlotClick} handleSlotDoubleClick={handleSlotDoubleClick}
+                               page={page} setPage={setPage} maxPage={MAX_PAGE}/>
                 </div>
                 <div className={'resizer-x'}/>
                 <div className="configurations">