+++ changelog.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+- Adding multiple pages in inventory, currently you can go up to 10 pages
+- Design change to allow more room for button configuration
+- You can now resize the page 
+
 # 1.0.5
 
 - Update color code